// apps/api-primary-node/src/auth/types.ts
export interface ResponseWithCookiesLike {
  cookie: (name: string, token: string, options: CookieOptions) => void,
  clearCookie: (name: string, options: CookieOptions) => void;
}
// ...
export interface AuthorizeUserAttributes {
  accessToken: string | undefined,
  refreshToken: string | undefined,
  userAgent: string | undefined,
}

export interface AuthorizeUserAttributesExtended extends AuthorizeUserAttributes {
  accessTokenSecret: string,
  accessTokenExpiry: string,
  refreshTokenSecret: string,
  refreshTokenExpiry: string,
}

export interface AuthorizeAPIKeyAttributes {
  apiKey: string,
}

export interface AuthorizeAPIKeyAttributesExtended extends AuthorizeAPIKeyAttributes {
  apiKeySecret: string,
  reqDomain: string,
  reqIP: string,
}

export interface AuthorizeUserResult {
  tenantId: string;
  tenantAccountId: string;
  accessToken?: string;
  refreshToken?: string;
}

// API Key types
export type ApiKeyType = "user" | "system";
export type ApiKeyScope = "read" | "write" | "admin" | "trigger";

export interface AuthorizeTenantResult {
  tenantId: string;
  keyType: ApiKeyType;
  scope: ApiKeyScope;
  userAccountId?: string;
}
// ...
export type AuthoriseUser = (attributes: AuthorizeUserAttributes, res: ResponseWithCookiesLike) => Promise<AuthorizeUserResult | null>;
export type AuthoriseAPIKey = (attributes: AuthorizeAPIKeyAttributes) => Promise<AuthorizeTenantResult | null>;

export interface CookieOptions {
  domain: string;
  path: string;
  httpOnly: boolean;
  secure: boolean;
  maxAge: number;
}

export type TenantType = 'user' | 'api' | 'service';

export type Tenant = {
  id: string;
} & (
  | {
      type: "user";
      userAccountId: string;
    }
  | {
      type: "api";
      keyType: ApiKeyType;
      scope: ApiKeyScope;
      userAccountId?: string;
    }
);

export interface RequestLike {
  cookies: {
    [key: string]: string;
  };
  headers: IncomingHttpHeaders;
}

export interface AuthRequest extends Request {
  tenant?: Tenant;
}